pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = ""
authors = [
    {name = "toreleon", email = "levietthang0512@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "websockets>=12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "pydantic[email]>=2.10.4",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "aiohttp>=3.11.11",
    "loguru>=0.7.3",
    "motor>=3.7.1",
    "gitpython>=3.1.43",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
